Store responses in to database continuously
This node script uses express and mssql packages mainly to store the api response in the database column on set frequency. No need to have postman runner or anything else except this code which itself calls API and prepares data which is helpful if user wants to query and analyze JSON response from api endpoint.